+++ packages/core/helper-plugin/lib/src/components/PaginationURLQuery/index.js
@@ -15,20 +15,14 @@
  */
 
 import React from 'react';
-import {
-  NextLink,
-  Pagination as PaginationCompo,
-  PreviousLink,
-  Dots,
-  PageLink,
-} from '@strapi/parts';
+import { NextLink, Pagination, PreviousLink, Dots, PageLink } from '@strapi/parts';
 import PropTypes from 'prop-types';
 import { useQueryParams } from '@strapi/helper-plugin';
 import { useLocation } from 'react-router-dom';
 import { useIntl } from 'react-intl';
 import { stringify } from 'qs';
 
-const Pagination = ({ pagination: { pageCount } }) => {
+const PaginationURLQuery = ({ pagination: { pageCount } }) => {
   const [{ query }] = useQueryParams();
   const activePage = parseInt(query?.page || '1', 10);
   const { pathname } = useLocation();
@@ -62,7 +56,7 @@ const Pagination = ({ pagination: { pageCount } }) => {
       });
 
     return (
-      <PaginationCompo activePage={activePage} pageCount={pageCount}>
+      <Pagination activePage={activePage} pageCount={pageCount}>
         <PreviousLink to={`${pathname}?${previousSearch}`}>
           {formatMessage({
             id: 'components.pagination.go-to-previous',
@@ -76,7 +70,7 @@ const Pagination = ({ pagination: { pageCount } }) => {
             defaultMessage: 'Go to next page',
           })}
         </NextLink>
-      </PaginationCompo>
+      </Pagination>
     );
   }
 
@@ -180,7 +174,7 @@ const Pagination = ({ pagination: { pageCount } }) => {
     : pageCount - activePage - 1;
 
   return (
-    <PaginationCompo activePage={activePage} pageCount={pageCount}>
+    <Pagination activePage={activePage} pageCount={pageCount}>
       <PreviousLink to={`${pathname}?${previousSearch}`}>
         {formatMessage({
           id: 'components.pagination.go-to-previous',
@@ -218,12 +212,12 @@ const Pagination = ({ pagination: { pageCount } }) => {
           defaultMessage: 'Go to next page',
         })}
       </NextLink>
-    </PaginationCompo>
+    </Pagination>
   );
 };
 
-Pagination.propTypes = {
+PaginationURLQuery.propTypes = {
   pagination: PropTypes.shape({ pageCount: PropTypes.number.isRequired }).isRequired,
 };
 
-export default Pagination;
+export default PaginationURLQuery;
